Finished Section Hiding, LaTeX Parsing, etc
* YOPP
YOPP stands for "YOPP Org-mode PHP Parser".  This project was started to
allow dynamic generation of HTML5/CSS3 files on the fly.  The final goal
of this project is to be able to conveniently write blogs with
org-mode.Here are some of the major new features that are implemented
today.
** DONE Section Hiding
Clicking on the header will hide the section under the header. The
currently algorithm uses a *stack* approach to balance out the tags. A
keyboard shortcut may be implemented in the future.
** DONE LaTeX Export
Using symbols such as =\(= and =\)= or =$$= and =$$= will interpret the
content inside these tags as LaTeX formula.
*For example:*
When \(a \ne 0\), there are two solutions to \(ax^2 + bx + c = 0\) and
they are
$$x = {-b \pm \sqrt{b^2-4ac} \over 2a}.$$
** DONE Intermediate Formatting
Using formatters such as =*=, =/=, and =+= will change the formatting.
*For Example:*
=*example*= will yield *example*.
=/example/= will yield /example/.
=+example+= will yield +example+
Many for the characters will be replaced by glyphs such as "double
quotes" and 'single quotes.' The characters in =<code>= tags, within
LaTeX operators, and within a URL are then converted back to their
"unglyphed" forms.
** STARTED List Handling
List items are started by starting a new line with =-=. Indenting the
list will create sublist or nested list.